refactor: #11 Organize Components File Folder
